requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.1.0
typing-extensions>=4.5.0
groq>=0.4.0
python-dotenv>=1.0.0
easyocr>=1.7.0
pillow>=10.0.0
opencv-python>=4.8.0
